% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\title{Checks whether the user has access to a resource}
\usage{
access(user, resource = c("App1", "App2", "App3"))
}
\arguments{
\item{user}{the user to check}

\item{resource}{one of
\itemize{
\item "App1"
\item "App2"
\item "App3"
}}
}
\value{
TRUE if the user has access, FALSE otherwise
}
\description{
This is for demonstration purposes, ideally this package is connected
to a database or other authentication system.
}
